requirements.txt
pulsar-client>=2.8.3,<2.9